4.21

+++ controllers/blogs.js
@@ -42,6 +42,16 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(400).json({ error: 'no blogs with given id'})
+  }
+
+  if (!request.token || blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({error: 'missing or invalid token'})
+  }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
